backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "google-genai>=1.65.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "qdrant-client>=1.17.0",
    "uvicorn[standard]>=0.41.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
